+++ hpy/devel/include/common/hpyfunc.h
