+++ hpy/devel/include/common/typeslots.h
